server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pyjwt>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
websockets>=11.0
openai>=1.0.0
